requirements.txt
django==5.1.6
djangorestframework==3.15.2
sentence-transformers==4.1.0
faiss-cpu==1.11.0
numpy==2.2.6 